First instance of Backbone.js modeling in Uskoci game. So far, only Hand is available via UskociCollection view.
