--- design/frontend/src/main/webapp/jquery/application.js
@@ -1,83 +1,0 @@
-$(function () {
-    "use strict";
-
-    var content = $('#content');
-    var input = $('#input');
-    var status = $('#status');
-    var myName = false;
-    var author = null;
-    var logged = false;
-    var socket = $.atmosphere;
-    var request = { url: document.location.toString() + 'rest/gamecontrol',
-                    contentType : "application/json",
-                    logLevel : 'debug',
-                    transport : 'long-polling' ,
-                    fallbackTransport: 'long-polling'};
-
-
-    request.onOpen = function(response) {
-        content.html($('<p>', { text: 'Atmosphere connected using ' + response.transport }));
-        input.removeAttr('disabled').focus();
-        status.text('Choose name:');
-    };
-
-    request.onMessage = function (response) {
-        var message = response.responseBody;
-        try {
-            var json = jQuery.parseJSON(message);
-        } catch (e) {
-            console.log('This doesn\'t look like a valid JSON: ', message.data);
-            return;
-        }
-
-        if (!logged) {
-            logged = true;
-            status.text(myName + ': ').css('color', 'blue');
-            input.removeAttr('disabled').focus();
-        } else {
-            input.removeAttr('disabled');
-
-            var me = json.author == author;
-            var date =  typeof(json.time) == 'string' ? parseInt(json.time) : json.time;
-            addMessage(json.author, json.text, me ? 'blue' : 'black', new Date(date));
-        }
-    };
-
-    request.onClose = function(response) {
-        logged = false;
-    }
-
-    request.onError = function(response) {
-        content.html($('<p>', { text: 'Sorry, but there\'s some problem with your '
-            + 'socket or the server is down' }));
-    };
-
-    var subSocket = socket.subscribe(request);
-
-    input.keydown(function(e) {
-        if (e.keyCode === 13) {
-            var msg = $(this).val();
-
-            // First message is always the author's name
-            if (author == null) {
-                author = msg;
-            }
-
-            subSocket.push(jQuery.stringifyJSON({ author: author, message: msg }));
-            $(this).val('');
-
-            input.attr('disabled', 'disabled');
-            if (myName === false) {
-                myName = msg;
-            }
-        }
-    });
-
-    function addMessage(author, message, color, datetime) {
-        content.append('<p><span style="color:' + color + '">' + author + '</span> @ ' +
-            + (datetime.getHours() < 10 ? '0' + datetime.getHours() : datetime.getHours()) + ':'
-            + (datetime.getMinutes() < 10 ? '0' + datetime.getMinutes() : datetime.getMinutes())
-            + ': ' + message + '</p>');
-    }
-});
-
