+++ design/frontend/src/main/webapp/js/fb-resource.js
